django_sample/openapi/clients/python/private/pyproject.toml
[tool.poetry]
name = "django_revolution_private"
version = "1.0.0 (v1)"
description = "A client library for accessing Django Revolution Test API"
authors = []
readme = "README.md"
packages = [
    {include = "django_revolution_private"},
]
include = ["CHANGELOG.md", "django_revolution_private/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
